requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
numpy==1.26.4
PyYAML==6.0.1
streamlit-authenticator==0.2.3
bcrypt==4.1.2
plotly>=5.20.0 